feat: add per-tab footer to UiTabs
