+++ src/components.d.ts
@@ -118,6 +118,7 @@ declare module 'vue' {
     SettingsSaveTab: typeof import('./components/settings/SaveTab.vue')['default']
     SettingsSettingsModal: typeof import('./components/settings/SettingsModal.vue')['default']
     SettingsShortcutsTab: typeof import('./components/settings/ShortcutsTab.vue')['default']
+    SettingsShortcutsTabFooter: typeof import('./components/settings/ShortcutsTabFooter.vue')['default']
     SettingsSupportTab: typeof import('./components/settings/SupportTab.vue')['default']
     ShlagemonDetail: typeof import('./components/shlagemon/Detail.vue')['default']
     ShlagemonDetailModal: typeof import('./components/shlagemon/DetailModal.vue')['default']
